Move actions, Adjust filepaths for app
