+++ app/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { grabMovies, displayMovies } from './containers/movieIndex/actions';
-import MovieIndex from './containers/movieIndex/MovieIndex';
+import { grabMovies, displayMovies } from './actions/MovieIndexActions';
+import MovieIndex from './containers/MovieIndex/MovieIndex';
 import FavoritesIndexContainer from './components/Favorites/FavoritesIndexContainer';
 import User from './components/User/User';
 import getRecentMovies from './utils/getRecentMovies';
