+++ app/components/User/User.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { logOut } from './actions';
+import { logOut } from '../../actions/UserActions';
 
 const User = ({ loginSuccessful, handleLogOut }) => {
